reuse device form, simplify realtime updates
